Import error bars from etterna and WF
This adds 3 timing visualization options:

-----

COLORFUL

The error bar from the Waterfall theme. It looks similar to the balance
meter known from the Tony Hawk's skating games. If you are a fan of
colors this is for you!

I made some modifications to it: No transparency for visual clarity,
always display all timing windows, some positioning tweaks and some
other changes that were necessary to make it work in the context of SL.

Thank you SteveReen for implementing it and giving us permission to use
it for SL!

-----

MONOCHROME

Another error bar implementation. If you appreciate the elegance of the
dark night with some occasional flashes of colored lights this is for
you!

It was made by prim for the spawncamping-wallhack theme and later
heavily modified by etterna devs, jordando and myself.

-----

TEXT

The early/late timing text from the Waterfall theme, adapted for SL. It
was also implemented by SteveReen.

(cherry picked from commit a216ddba6e411569c20cf23de8f22b809020b8d9)

# Conflicts:
#	Scripts/SL-PlayerProfiles.lua
#	metrics.ini