ensure DateTime footer appears in CourseMode Eval
Commit e26408955220e706acc6145528adad9842fa5e7c added support for
showing the date and time in a BitmapText actor at the bottom of SL's
evaluation screens.

It did so by adding code to ScreenSystemLayer overlay, and then by
toggling visibility of the BitmapText actor based on the current
screen's name.  This worked, though the implementation accidentally left
"ScreenEvaluationNonstop" out of the if-else chain that checked screen
name, causing the Date and Time to not appear at the bottom of
CourseMode's evaluation screen.

This commit accomplishes a few things:

1. Extracts the BitampText and accompanying ActorFrame out of
ScrenSystemLayer overlay and moves them to ScreenEvaluation decorations

Since the date and time only appear in evaluation screens, it didn't
seem necessary to have their code in a global overlay.

2. Removes code to toggle visibility based on current screen name. It's
not needed any longer. This is a beneficial result of change 1.

3. Adjusts the y-offset code that relies on current screen name to more
generally accommodate screens that inherit from ScreenEvaluation, as
that's when this file is going to be loaded.  In SL right now, this
includes ScreenEvalationStage, ScreenEvaluationNonstop, and
ScreenEvaluationSummary.

4. Changes the scope of the update function be local to this file.
There didn't seem to be a need to add it to Lua's global namespace.

5. Added inline comments providing context and explanation.

(cherry picked from commit d3994fbebc31cc03e62aeb610e4df8ae7b10595e)