display additional precision for mean timing error
Issue #253 on GitHub requested that ScreenEval/Pane4 be extended to show
1 decimal place of precision for mean timing error, with that idea that
"4.6ms is actually a lot better than 5.4ms but they would both
[currently] show up as 5ms mean timing error."

teejusb responded:
"I think a nice approach to this is to just add the significant digit
only in cases where the metrics are < 10ms. This way we mostly preserve
the space utilization (we add a period) and get additional info when it
seems to be worthwhile."

This commit implements the request using teejusb's suggested approach.

(cherry picked from commit 0ccd6ef0255a254cfcb98b46088bc7df9649eeec)